requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
ray[default]==2.46.0
pandas==2.3.0
numpy==2.2.6
scikit-learn==1.6.1
xgboost==3.0.2
mlflow==2.22.1
pyyaml==6.0.2
joblib==1.5.1
opentelemetry-api==1.34.0
opentelemetry-sdk==1.34.0
opentelemetry-exporter-otlp==1.34.0
opentelemetry-instrumentation-fastapi==0.55b0
